EZP-30461: Implemented Generic Field Type (CR)
