--- eZ/Publish/Core/FieldType/Generic/BaseValue.php
@@ -1,16 +1,0 @@
-<?php
-
-/**
- * @copyright Copyright (C) eZ Systems AS. All rights reserved.
- * @license For full copyright and license information view LICENSE file distributed with this source code.
- */
-declare(strict_types=1);
-
-namespace eZ\Publish\Core\FieldType\Generic;
-
-use eZ\Publish\Core\FieldType\Value;
-use eZ\Publish\Core\FieldType\Generic\Value as ValueInterface;
-
-abstract class BaseValue extends Value implements ValueInterface
-{
-}
