+++ eZ/Publish/Core/settings/fieldtype_services.yml
@@ -54,3 +54,4 @@ services:
             $locationService: '@ezpublish.api.service.location'
             $contentTypeService: '@ezpublish.api.service.content_type'
             $mappings: '$fieldtypes.ezimageasset.mappings$'
+
