+++ eZ/Publish/Core/settings/storage_engines/legacy/field_value_converters.yml
@@ -33,10 +33,9 @@ services:
         class: "%ezpublish.persistence.legacy.field_value_converter.registry.class%"
         lazy: true
 
-    ezpublish.fieldType.ezgeneric.converter:
-        class: eZ\Publish\Core\Persistence\Legacy\Content\FieldValue\Converter\GenericConverter
+    eZ\Publish\Core\Persistence\Legacy\Content\FieldValue\Converter\SerializableConverter:
         arguments:
-            $serializer: '@eZ\Publish\Core\FieldType\Generic\ValueSerializerInterface'
+            $serializer: '@eZ\Publish\SPI\FieldType\ValueSerializerInterface'
 
     ezpublish.fieldType.ezauthor.converter:
         class: "%ezpublish.fieldType.ezauthor.converter.class%"
